Add content, update structure
