+++ .vscode/settings.json
@@ -4,5 +4,6 @@
   "editor.formatOnType": false, // required
   "editor.formatOnSave": true, // optional
   "editor.formatOnSaveMode": "file", // required to format on save
-  "files.autoSave": "onFocusChange" // optional but recommended
+  "files.autoSave": "onFocusChange",
+  "githubPullRequests.ignoredPullRequestBranches": ["main"] // optional but recommended
 }
