+++ package.json
@@ -38,6 +38,7 @@
     "@vueuse/motion": "2.0.0-beta.12",
     "popmotion": "^11.0.5",
     "sharp": "^0.31.0",
+    "swiper": "^8.4.2",
     "vue": "^3.2.39",
     "vue-i18n": "^9.2.2",
     "vue-router": "^4.1.5"
