+++ public/data/HomePage.json
@@ -1,6 +1,7 @@
 {
   "news": [
     {
+      "date": "2022-09-01",
       "zh-CN": {
         "title": "题目一",
         "abstract": "摘要一",
@@ -15,6 +16,7 @@
       }
     },
     {
+      "date": "2022-09-01",
       "zh-CN": {
         "title": "题目二",
         "abstract": "摘要二",
@@ -29,6 +31,7 @@
       }
     },
     {
+      "date": "2022-09-01",
       "zh-CN": {
         "title": "题目三",
         "abstract": "摘要三",
@@ -38,6 +41,21 @@
       "en-US": {
         "title": "Title Three",
         "abstract": "Abstract Three",
+        "color": "#B5C2A1",
+        "link": "/articles/YYYY/MM/dd/some-article-name"
+      }
+    },
+    {
+      "date": "2022-09-01",
+      "zh-CN": {
+        "title": "题目四",
+        "abstract": "摘要四",
+        "color": "#B5C2A1",
+        "link": "/articles/YYYY/MM/dd/some-article-name"
+      },
+      "en-US": {
+        "title": "Title Four",
+        "abstract": "Abstract Four",
         "color": "#B5C2A1",
         "link": "/articles/YYYY/MM/dd/some-article-name"
       }
