+++ src/components/DesktopHomePage/SecondSection.vue
@@ -1,8 +1,8 @@
 <template>
-  <section id="secondSection" class="p-y-90px">
+  <div class="p-y-90px">
     <div class="section">
       <div class="flex">
-        <div class="flex-1">
+        <div class="flex-1 mr-4">
           <h2 class="color-[var(--standard-blue)] text-9 m-t-0">
             {{ $t('HomePage.SecondSection.Title') }}
           </h2>
@@ -11,7 +11,7 @@
             <ReadMore to="/about#overview" />
           </div>
         </div>
-        <div class="flex-1 flex justify-center items-center">
+        <div class="ml-4 flex-1 flex justify-center items-center">
           <video class="w-480px h-270px" controls>
             <source :src="vidOverview" type="video/mp4" />
             Your browser does not support the video tag.
@@ -19,7 +19,7 @@
         </div>
       </div>
     </div>
-  </section>
+  </div>
 </template>
 
 <script setup>
